+++ poetry/masonry/builders/sdist.py
@@ -119,6 +119,10 @@ class SdistBuilder(Builder):
                 'build(setup_kwargs)'
             ]
 
+        if self._module.is_in_src():
+            before.append("package_dir = \\\n{}\n".format(pformat({'': 'src'})))
+            extra.append("'package_dir': package_dir,")
+
         if self._module.is_package():
             packages, package_data = self.find_packages(
                 self._module.path.as_posix()